refactor: components refactoring
Creating folders for components to bring code clarity
