+++ app/page.tsx
@@ -3,7 +3,7 @@ import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
 import Image from "next/image";
 import Link from "next/link";
-import { useCallback, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { whyte } from "./fonts";
 import {
   Accordion,
@@ -12,8 +12,10 @@ import {
   AccordionTrigger,
 } from "@/components/ui/accordion"
 import { Separator } from "@/components/ui/separator";
-import { toast, useToast } from "@/components/ui/use-toast";
+import { toast } from "@/components/ui/use-toast";
 import { Toaster } from "@/components/ui/toaster";
+import ReliabilityCertification from "@/components/Certifications/ReliabilityCertification";
+import SecurityCertification from "@/components/Certifications/SecurityCertification";
 
 export default function Home() {
   const [error, setError] = useState<string | null>(null);
@@ -35,20 +37,7 @@ export default function Home() {
     }
   }, [error]);
 
-  const [highlight, setHighlight] = useState<boolean>(false);
-  const [openPopover, setOpenPopover] = useState<boolean>(false);
-  const [typing, setTyping] = useState<boolean>(false);
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      // Your code here
-      setHighlight(true);
-      setOpenPopover(true);
-      console.log('highlighting');
-    }, 2000); // Delay of 2 seconds
 
-    // Clear timeout if the component is unmounted
-    return () => clearTimeout(timer);
-  }, []);
   return (
     <div className="min-h-full mb-24 w-full overflow-visible">
       <Toaster />
@@ -100,82 +89,8 @@ export default function Home() {
             </div>
           </h3>
           <div className="self-center justify-center flex overflow-visible flex-wrap sm:justify-start">
-            <div className="flex w-36 h-36 sm:w-64 sm:h-64 overflow-visible">
-              <svg viewBox="0 0 825 746" fill="none" xmlns="http://www.w3.org/2000/svg" className=" overflow-visible">
-                <g filter="url(#filter0_d_421_1162)">
-                  <path d="M329.126 60.3472C356.016 32.3351 400.815 32.3351 427.705 60.3472L433.004 65.8681C449.747 83.3094 474.46 90.604 497.99 85.0497L505.306 83.3227C542.941 74.439 580.441 98.5309 588.005 136.453L589.767 145.283C594.455 168.786 611.115 188.113 633.669 196.216L641.906 199.176C678.021 212.151 696.369 252.317 682.532 288.11L678.932 297.422C670.354 319.614 673.937 344.67 688.393 363.568L694.414 371.44C717.678 401.853 711.409 445.446 680.517 468.071L672.891 473.656C653.644 487.752 643.144 510.867 645.19 534.635L646.017 544.246C649.319 582.62 620.267 616.138 581.813 618.318L573.795 618.773C549.749 620.136 528.197 634.06 517.069 655.42L513.253 662.746C495.333 697.145 452.406 709.744 418.738 690.488L412.336 686.827C391.32 674.807 365.511 674.807 344.495 686.827L338.093 690.488C304.425 709.744 261.498 697.145 243.578 662.746L239.762 655.42C228.634 634.06 207.082 620.136 183.036 618.773L175.018 618.318C136.564 616.138 107.512 582.62 110.814 544.246L111.641 534.635C113.687 510.867 103.187 487.752 83.9399 473.656L76.3141 468.071C45.4218 445.446 39.1526 401.853 62.417 371.44L68.4384 363.568C82.8936 344.67 86.4771 319.614 77.8984 297.422L74.2987 288.11C60.4619 252.317 78.8102 212.151 114.925 199.176L123.162 196.216C145.716 188.113 162.376 168.785 167.064 145.283L168.825 136.453C176.39 98.5309 213.89 74.439 251.525 83.3227L258.841 85.0497C282.371 90.604 307.084 83.3094 323.827 65.868L329.126 60.3472Z" fill="url(#paint0_linear_421_1162)" />
-                  <path d="M329.126 60.3472C356.016 32.3351 400.815 32.3351 427.705 60.3472L433.004 65.8681C449.747 83.3094 474.46 90.604 497.99 85.0497L505.306 83.3227C542.941 74.439 580.441 98.5309 588.005 136.453L589.767 145.283C594.455 168.786 611.115 188.113 633.669 196.216L641.906 199.176C678.021 212.151 696.369 252.317 682.532 288.11L678.932 297.422C670.354 319.614 673.937 344.67 688.393 363.568L694.414 371.44C717.678 401.853 711.409 445.446 680.517 468.071L672.891 473.656C653.644 487.752 643.144 510.867 645.19 534.635L646.017 544.246C649.319 582.62 620.267 616.138 581.813 618.318L573.795 618.773C549.749 620.136 528.197 634.06 517.069 655.42L513.253 662.746C495.333 697.145 452.406 709.744 418.738 690.488L412.336 686.827C391.32 674.807 365.511 674.807 344.495 686.827L338.093 690.488C304.425 709.744 261.498 697.145 243.578 662.746L239.762 655.42C228.634 634.06 207.082 620.136 183.036 618.773L175.018 618.318C136.564 616.138 107.512 582.62 110.814 544.246L111.641 534.635C113.687 510.867 103.187 487.752 83.9399 473.656L76.3141 468.071C45.4218 445.446 39.1526 401.853 62.417 371.44L68.4384 363.568C82.8936 344.67 86.4771 319.614 77.8984 297.422L74.2987 288.11C60.4619 252.317 78.8102 212.151 114.925 199.176L123.162 196.216C145.716 188.113 162.376 168.785 167.064 145.283L168.825 136.453C176.39 98.5309 213.89 74.439 251.525 83.3227L258.841 85.0497C282.371 90.604 307.084 83.3094 323.827 65.868L329.126 60.3472Z" fill="url(#paint1_linear_421_1162)" fillOpacity="0.1" />
-                  <path d="M129.935 372.236C129.935 235.005 241.182 123.757 378.413 123.757C515.644 123.757 626.892 235.005 626.892 372.236V377.346C626.892 514.577 515.644 625.825 378.413 625.825C241.182 625.825 129.935 514.577 129.935 377.346V372.236Z" fill="black" fillOpacity="0.04" />
-                  <path d="M155.445 372.236C155.445 249.094 255.271 149.267 378.413 149.267C501.556 149.267 601.382 249.094 601.382 372.236V377.346C601.382 500.489 501.556 600.315 378.413 600.315C255.271 600.315 155.445 500.488 155.445 377.346V372.236Z" stroke="white" strokeOpacity="0.3" strokeWidth="51.0205" />
-                  <path d="M630.253 377.07C630.253 516.156 517.501 628.907 378.415 628.907C239.329 628.907 126.578 516.156 126.578 377.07C126.578 237.984 239.329 125.233 378.415 125.233C517.501 125.233 630.253 237.984 630.253 377.07ZM178.425 377.07C178.425 487.521 267.964 577.06 378.415 577.06C488.867 577.06 578.405 487.521 578.405 377.07C578.405 266.618 488.867 177.08 378.415 177.08C267.964 177.08 178.425 266.618 178.425 377.07Z" fill="url(#paint2_linear_421_1162)" fillOpacity="0.5" />
-                  <path d="M264.793 388.864C261.497 388.864 258.649 388.448 256.249 387.616C253.849 386.752 251.945 385.552 250.537 384.016C249.161 382.48 248.297 380.656 247.945 378.544L259.561 376.192C259.721 377.12 260.025 377.936 260.473 378.64C260.921 379.312 261.513 379.856 262.249 380.272C263.017 380.656 263.929 380.848 264.985 380.848C266.489 380.848 267.593 380.608 268.297 380.128C269.001 379.616 269.353 378.944 269.353 378.112C269.353 377.408 269.065 376.816 268.489 376.336C267.945 375.856 267.129 375.552 266.041 375.424L261.865 374.944C258.921 374.592 256.553 374 254.761 373.168C252.969 372.304 251.657 371.168 250.825 369.76C249.993 368.352 249.577 366.576 249.577 364.432C249.577 362.32 250.217 360.448 251.497 358.816C252.777 357.152 254.585 355.856 256.921 354.928C259.257 354 261.977 353.536 265.081 353.536C268.057 353.536 270.665 353.92 272.905 354.688C275.177 355.424 277.001 356.512 278.377 357.952C279.753 359.392 280.649 361.104 281.065 363.088L269.401 365.248C269.209 364.512 268.889 363.856 268.441 363.28C268.025 362.704 267.481 362.256 266.809 361.936C266.169 361.616 265.433 361.456 264.601 361.456C263.833 361.456 263.161 361.568 262.585 361.792C262.041 362.016 261.625 362.32 261.337 362.704C261.049 363.088 260.905 363.52 260.905 364C260.905 364.736 261.177 365.328 261.721 365.776C262.297 366.192 263.161 366.464 264.313 366.592L268.585 367.072C272.649 367.488 275.785 368.624 277.993 370.48C280.233 372.304 281.353 374.704 281.353 377.68C281.353 379.952 280.713 381.936 279.433 383.632C278.153 385.296 276.265 386.592 273.769 387.52C271.305 388.416 268.313 388.864 264.793 388.864ZM285.304 388V354.4H312.232V362.992H296.344V367.168H309.064V374.896H296.344V379.36H312.232V388H285.304ZM333.382 388.864C329.574 388.864 326.262 388.16 323.446 386.752C320.662 385.344 318.518 383.328 317.014 380.704C315.51 378.08 314.758 374.992 314.758 371.44C314.758 367.984 315.526 364.912 317.062 362.224C318.598 359.504 320.774 357.376 323.59 355.84C326.438 354.304 329.734 353.536 333.478 353.536C336.358 353.536 338.998 354 341.398 354.928C343.798 355.824 345.814 357.12 347.446 358.816C349.11 360.512 350.294 362.496 350.998 364.768L339.382 368.32C339.062 367.008 338.646 365.92 338.134 365.056C337.622 364.192 336.95 363.536 336.118 363.088C335.318 362.64 334.358 362.416 333.238 362.416C331.958 362.416 330.854 362.768 329.926 363.472C329.03 364.144 328.326 365.168 327.814 366.544C327.334 367.888 327.094 369.52 327.094 371.44C327.094 373.136 327.334 374.64 327.814 375.952C328.326 377.232 329.046 378.224 329.974 378.928C330.902 379.632 331.99 379.984 333.238 379.984C334.806 379.984 336.086 379.504 337.078 378.544C338.102 377.584 338.918 376.08 339.526 374.032L351.046 377.92C350.342 380.064 349.158 381.968 347.494 383.632C345.83 385.264 343.782 386.544 341.35 387.472C338.918 388.4 336.262 388.864 333.382 388.864ZM370.338 388.864C366.69 388.864 363.634 388.32 361.17 387.232C358.738 386.112 356.914 384.432 355.698 382.192C354.482 379.92 353.874 377.04 353.874 373.552V354.4H365.346V372.352C365.346 374.912 365.73 376.752 366.498 377.872C367.266 378.992 368.53 379.552 370.29 379.552C372.018 379.552 373.266 378.992 374.034 377.872C374.834 376.752 375.234 374.912 375.234 372.352V354.4H386.754V373.552C386.754 377.04 386.146 379.92 384.93 382.192C383.714 384.432 381.874 386.112 379.41 387.232C376.978 388.32 373.954 388.864 370.338 388.864ZM391.663 388V354.4H410.911C413.503 354.4 415.711 354.752 417.535 355.456C419.391 356.16 420.783 357.168 421.711 358.48C422.639 359.76 423.103 361.248 423.103 362.944C423.103 365.056 422.495 366.688 421.279 367.84C420.095 368.96 418.623 369.712 416.863 370.096C415.135 370.48 413.199 370.672 411.055 370.672V374.032C413.583 374.032 415.599 374.256 417.103 374.704C418.607 375.12 419.807 375.872 420.703 376.96C421.631 378.048 422.351 379.616 422.863 381.664L424.399 388H412.351L410.671 379.84C410.479 378.912 410.239 378.192 409.951 377.68C409.663 377.136 409.263 376.752 408.751 376.528C408.271 376.272 407.615 376.144 406.783 376.144H402.895V388H391.663ZM407.887 368.848C408.975 368.848 409.839 368.576 410.479 368.032C411.151 367.456 411.487 366.72 411.487 365.824C411.487 364.896 411.167 364.16 410.527 363.616C409.887 363.072 409.007 362.8 407.887 362.8H402.895V368.848H407.887ZM427.945 388V354.4H439.561V388H427.945ZM452.787 388V362.512H442.803V354.4H474.387V362.512H464.355V388H452.787ZM486.514 388V375.616L474.514 354.4H486.802L488.626 357.952C489.618 359.904 490.242 361.872 490.498 363.856C490.786 365.808 490.898 368.08 490.834 370.672H494.386C494.322 368.208 494.466 365.888 494.818 363.712C495.17 361.504 495.826 359.408 496.786 357.424L498.37 354.4H510.13L498.082 375.616V388H486.514Z" fill="white" />
-                  <path d="M254.605 337L265.245 309H273.405L283.925 337H277.005L271.085 319.56C270.792 318.68 270.592 317.72 270.485 316.68C270.405 315.613 270.365 314.2 270.365 312.44H268.125C268.125 314.28 268.072 315.707 267.965 316.72C267.885 317.733 267.685 318.693 267.365 319.6L261.325 337H254.605ZM261.005 330.52L262.765 325.4H275.725L277.485 330.52H261.005ZM287.374 337V309H300.054C302.054 309 303.814 309.373 305.334 310.12C306.88 310.84 308.08 311.893 308.934 313.28C309.787 314.667 310.214 316.267 310.214 318.08C310.214 319.947 309.787 321.573 308.934 322.96C308.08 324.32 306.88 325.373 305.334 326.12C303.814 326.867 302.067 327.24 300.094 327.24H293.654V337H287.374ZM298.934 322.12C299.867 322.12 300.694 321.96 301.414 321.64C302.134 321.32 302.694 320.867 303.094 320.28C303.52 319.667 303.734 318.933 303.734 318.08C303.734 317.307 303.52 316.627 303.094 316.04C302.694 315.427 302.134 314.96 301.414 314.64C300.694 314.32 299.867 314.16 298.934 314.16H293.654V322.12H298.934ZM314.405 337V309H327.085C329.085 309 330.845 309.373 332.365 310.12C333.912 310.84 335.112 311.893 335.965 313.28C336.818 314.667 337.245 316.267 337.245 318.08C337.245 319.947 336.818 321.573 335.965 322.96C335.112 324.32 333.912 325.373 332.365 326.12C330.845 326.867 329.098 327.24 327.125 327.24H320.685V337H314.405ZM325.965 322.12C326.898 322.12 327.725 321.96 328.445 321.64C329.165 321.32 329.725 320.867 330.125 320.28C330.552 319.667 330.765 318.933 330.765 318.08C330.765 317.307 330.552 316.627 330.125 316.04C329.725 315.427 329.165 314.96 328.445 314.64C327.725 314.32 326.898 314.16 325.965 314.16H320.685V322.12H325.965ZM340.676 324.88V319.6H355.076V324.88H340.676ZM360.889 337V309H373.569C375.569 309 377.329 309.373 378.849 310.12C380.396 310.84 381.596 311.893 382.449 313.28C383.303 314.667 383.729 316.267 383.729 318.08C383.729 319.947 383.303 321.573 382.449 322.96C381.596 324.32 380.396 325.373 378.849 326.12C377.329 326.867 375.583 327.24 373.609 327.24H367.169V337H360.889ZM372.449 322.12C373.383 322.12 374.209 321.96 374.929 321.64C375.649 321.32 376.209 320.867 376.609 320.28C377.036 319.667 377.249 318.933 377.249 318.08C377.249 317.307 377.036 316.627 376.609 316.04C376.209 315.427 375.649 314.96 374.929 314.64C374.209 314.32 373.383 314.16 372.449 314.16H367.169V322.12H372.449ZM387.921 337V309H401.881C403.881 309 405.601 309.307 407.041 309.92C408.481 310.507 409.574 311.36 410.321 312.48C411.094 313.6 411.481 314.893 411.481 316.36C411.481 318.2 411.027 319.587 410.121 320.52C409.241 321.453 408.187 322.067 406.961 322.36C405.761 322.653 404.361 322.84 402.761 322.92V325C404.441 325.027 405.814 325.187 406.881 325.48C407.947 325.773 408.814 326.307 409.481 327.08C410.147 327.827 410.601 328.92 410.841 330.36L412.001 337H405.321L404.441 330.76C404.281 329.693 404.001 328.853 403.601 328.24C403.227 327.6 402.681 327.133 401.961 326.84C401.241 326.547 400.307 326.4 399.161 326.4H394.201V337H387.921ZM400.681 321.4C401.987 321.4 403.027 321.093 403.801 320.48C404.601 319.867 405.001 318.96 405.001 317.76C405.001 316.587 404.601 315.693 403.801 315.08C403.027 314.467 401.987 314.16 400.681 314.16H394.201V321.4H400.681ZM430.149 337.64C427.029 337.64 424.335 337.027 422.069 335.8C419.802 334.547 418.069 332.813 416.869 330.6C415.695 328.387 415.109 325.88 415.109 323.08C415.109 320.28 415.709 317.773 416.909 315.56C418.135 313.32 419.869 311.56 422.109 310.28C424.375 309 427.055 308.36 430.149 308.36C433.215 308.36 435.869 309 438.109 310.28C440.375 311.56 442.109 313.32 443.309 315.56C444.535 317.773 445.149 320.28 445.149 323.08C445.149 325.88 444.549 328.387 443.349 330.6C442.175 332.813 440.455 334.547 438.189 335.8C435.922 337.027 433.242 337.64 430.149 337.64ZM430.189 332.48C431.895 332.48 433.389 332.08 434.669 331.28C435.949 330.453 436.922 329.333 437.589 327.92C438.255 326.48 438.589 324.853 438.589 323.04C438.589 321.2 438.255 319.56 437.589 318.12C436.922 316.68 435.949 315.56 434.669 314.76C433.389 313.933 431.895 313.52 430.189 313.52C428.429 313.52 426.909 313.933 425.629 314.76C424.349 315.56 423.375 316.68 422.709 318.12C422.042 319.56 421.709 321.2 421.709 323.04C421.709 324.853 422.042 326.48 422.709 327.92C423.375 329.333 424.349 330.453 425.629 331.28C426.909 332.08 428.429 332.48 430.189 332.48ZM456.206 337L446.526 309H453.526L458.126 323.92C458.633 325.6 458.993 327.107 459.206 328.44C459.419 329.747 459.526 331.133 459.526 332.6H461.766C461.766 331.16 461.873 329.787 462.086 328.48C462.299 327.147 462.673 325.627 463.206 323.92L467.846 309H474.726L465.086 337H456.206ZM478.468 337V309H499.708V314.16H484.748V320.08H496.108V325.12H484.748V331.84H499.708V337H478.468Z" fill="white" />
-                </g>
-                <defs>
-                  <filter id="filter0_d_421_1162" x="-96.2008" y="-29.7442" width="949.233" height="949.233" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
-                    <feFlood floodOpacity="0" result="BackgroundImageFix" />
-                    <feColorMatrix in="SourceAlpha" type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha" />
-                    <feOffset dy="77.4884" />
-                    <feGaussianBlur stdDeviation="58.1163" />
-                    <feComposite in2="hardAlpha" operator="out" />
-                    <feColorMatrix type="matrix" values="0 0 0 0 1 0 0 0 0 0.25098 0 0 0 0 0.0117647 0 0 0 0.25 0" />
-                    <feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow_421_1162" />
-                    <feBlend mode="normal" in="SourceGraphic" in2="effect1_dropShadow_421_1162" result="shape" />
-                  </filter>
-                  <linearGradient id="paint0_linear_421_1162" x1="378.415" y1="9.00021" x2="378.415" y2="725.768" gradientUnits="userSpaceOnUse">
-                    <stop stopColor="#FF4003" />
-                    <stop offset="1" stopColor="#992602" />
-                  </linearGradient>
-                  <linearGradient id="paint1_linear_421_1162" x1="-103.062" y1="-46.9972" x2="441.42" y2="924.853" gradientUnits="userSpaceOnUse">
-                    <stop stopColor="white" />
-                    <stop offset="1" stopColor="white" stopOpacity="0" />
-                  </linearGradient>
-                  <linearGradient id="paint2_linear_421_1162" x1="378.415" y1="125.233" x2="378.415" y2="628.907" gradientUnits="userSpaceOnUse">
-                    <stop stopColor="white" />
-                    <stop offset="0.0001" stopColor="white" stopOpacity="0.81" />
-                    <stop offset="1" stopColor="white" stopOpacity="0" />
-                  </linearGradient>
-                </defs>
-              </svg>
-            </div>
-            <div className="flex w-36 h-36 sm:w-64 sm:h-64">
-              <svg viewBox="0 0 826 746" fill="none" xmlns="http://www.w3.org/2000/svg" className=" overflow-visible">
-                <g filter="url(#filter0_d_421_1186)">
-                  <path d="M397.894 60.3472C424.783 32.3351 469.583 32.3351 496.472 60.3472L501.772 65.8681C518.514 83.3094 543.228 90.604 566.758 85.0497L574.074 83.3227C611.708 74.439 649.208 98.5309 656.773 136.453L658.535 145.283C663.223 168.786 679.883 188.113 702.437 196.216L710.673 199.176C746.788 212.151 765.137 252.317 751.3 288.11L747.7 297.422C739.121 319.614 742.705 344.67 757.16 363.568L763.182 371.44C786.446 401.853 780.177 445.446 749.284 468.071L741.659 473.656C722.412 487.752 711.911 510.867 713.957 534.635L714.784 544.246C718.087 582.62 689.035 616.138 650.581 618.318L642.563 618.773C618.516 620.136 596.964 634.06 585.837 655.42L582.02 662.746C564.101 697.145 521.174 709.744 487.506 690.488L481.103 686.827C460.087 674.807 434.279 674.807 413.263 686.827L406.86 690.488C373.192 709.744 330.265 697.145 312.346 662.746L308.529 655.42C297.402 634.06 275.85 620.136 251.803 618.773L243.785 618.318C205.331 616.138 176.279 582.62 179.582 544.246L180.409 534.635C182.455 510.867 171.954 487.752 152.707 473.656L145.082 468.071C114.189 445.446 107.92 401.853 131.185 371.44L137.206 363.568C151.661 344.67 155.245 319.614 146.666 297.422L143.066 288.11C129.229 252.317 147.578 212.151 183.693 199.176L191.929 196.216C214.484 188.113 231.143 168.785 235.832 145.283L237.593 136.453C245.158 98.5309 282.658 74.439 320.292 83.3227L327.608 85.0497C351.138 90.604 375.852 83.3094 392.594 65.868L397.894 60.3472Z" fill="url(#paint0_linear_421_1186)" />
-                  <path d="M397.894 60.3472C424.783 32.3351 469.583 32.3351 496.472 60.3472L501.772 65.8681C518.514 83.3094 543.228 90.604 566.758 85.0497L574.074 83.3227C611.708 74.439 649.208 98.5309 656.773 136.453L658.535 145.283C663.223 168.786 679.883 188.113 702.437 196.216L710.673 199.176C746.788 212.151 765.137 252.317 751.3 288.11L747.7 297.422C739.121 319.614 742.705 344.67 757.16 363.568L763.182 371.44C786.446 401.853 780.177 445.446 749.284 468.071L741.659 473.656C722.412 487.752 711.911 510.867 713.957 534.635L714.784 544.246C718.087 582.62 689.035 616.138 650.581 618.318L642.563 618.773C618.516 620.136 596.964 634.06 585.837 655.42L582.02 662.746C564.101 697.145 521.174 709.744 487.506 690.488L481.103 686.827C460.087 674.807 434.279 674.807 413.263 686.827L406.86 690.488C373.192 709.744 330.265 697.145 312.346 662.746L308.529 655.42C297.402 634.06 275.85 620.136 251.803 618.773L243.785 618.318C205.331 616.138 176.279 582.62 179.582 544.246L180.409 534.635C182.455 510.867 171.954 487.752 152.707 473.656L145.082 468.071C114.189 445.446 107.92 401.853 131.185 371.44L137.206 363.568C151.661 344.67 155.245 319.614 146.666 297.422L143.066 288.11C129.229 252.317 147.578 212.151 183.693 199.176L191.929 196.216C214.484 188.113 231.143 168.785 235.832 145.283L237.593 136.453C245.158 98.5309 282.658 74.439 320.292 83.3227L327.608 85.0497C351.138 90.604 375.852 83.3094 392.594 65.868L397.894 60.3472Z" fill="url(#paint1_linear_421_1186)" fillOpacity="0.1" />
-                  <path d="M198.702 372.236C198.702 235.005 309.95 123.757 447.181 123.757C584.412 123.757 695.66 235.005 695.66 372.236V377.346C695.66 514.577 584.412 625.825 447.181 625.825C309.95 625.825 198.702 514.577 198.702 377.346V372.236Z" fill="black" fillOpacity="0.04" />
-                  <path d="M224.212 372.236C224.212 249.094 324.039 149.267 447.181 149.267C570.323 149.267 670.149 249.094 670.149 372.236V377.346C670.149 500.489 570.323 600.315 447.181 600.315C324.039 600.315 224.212 500.488 224.212 377.346V372.236Z" stroke="white" strokeOpacity="0.3" strokeWidth="51.0205" />
-                  <path d="M699.02 377.07C699.02 516.156 586.269 628.907 447.183 628.907C308.097 628.907 195.346 516.156 195.346 377.07C195.346 237.984 308.097 125.233 447.183 125.233C586.269 125.233 699.02 237.984 699.02 377.07ZM247.193 377.07C247.193 487.521 336.731 577.06 447.183 577.06C557.634 577.06 647.173 487.521 647.173 377.07C647.173 266.618 557.634 177.08 447.183 177.08C336.731 177.08 247.193 266.618 247.193 377.07Z" fill="url(#paint2_linear_421_1186)" fillOpacity="0.5" />
-                  <path d="M290.415 388.233V354.633H309.663C312.255 354.633 314.463 354.985 316.287 355.689C318.143 356.393 319.535 357.401 320.463 358.713C321.391 359.993 321.855 361.481 321.855 363.177C321.855 365.289 321.247 366.921 320.031 368.073C318.847 369.193 317.375 369.945 315.615 370.329C313.887 370.713 311.951 370.905 309.807 370.905V374.265C312.335 374.265 314.351 374.489 315.855 374.937C317.359 375.353 318.559 376.105 319.455 377.193C320.383 378.281 321.103 379.849 321.615 381.897L323.151 388.233H311.103L309.423 380.073C309.231 379.145 308.991 378.425 308.703 377.913C308.415 377.369 308.015 376.985 307.503 376.761C307.023 376.505 306.367 376.377 305.535 376.377H301.647V388.233H290.415ZM306.639 369.081C307.727 369.081 308.591 368.809 309.231 368.265C309.903 367.689 310.239 366.953 310.239 366.057C310.239 365.129 309.919 364.393 309.279 363.849C308.639 363.305 307.759 363.033 306.639 363.033H301.647V369.081H306.639ZM326.697 388.233V354.633H353.625V363.225H337.737V367.401H350.457V375.129H337.737V379.593H353.625V388.233H326.697ZM358.103 388.233V354.633H369.719V379.353H385.559V388.233H358.103ZM388.806 388.233V354.633H400.422V388.233H388.806ZM402.848 388.233L414.08 354.633H429.584L440.912 388.233H428.672L424.208 372.441C423.792 370.841 423.536 369.017 423.44 366.969C423.376 364.921 423.392 362.569 423.488 359.913H419.84C419.936 362.313 419.904 364.585 419.744 366.729C419.584 368.873 419.264 370.777 418.784 372.441L414.128 388.233H402.848ZM410.912 382.473L413.84 374.361H429.392L432.368 382.473H410.912ZM443.228 388.233V354.633H460.94C463.34 354.633 465.388 354.937 467.084 355.545C468.812 356.153 470.108 357.017 470.972 358.137C471.836 359.257 472.268 360.537 472.268 361.977C472.268 364.249 471.452 366.009 469.82 367.257C468.188 368.473 465.74 369.113 462.476 369.177V372.489C464.908 372.489 466.94 372.761 468.572 373.305C470.236 373.849 471.484 374.697 472.316 375.849C473.148 376.969 473.564 378.393 473.564 380.121C473.564 381.817 473.068 383.273 472.076 384.489C471.116 385.705 469.692 386.633 467.804 387.273C465.948 387.913 463.724 388.233 461.132 388.233H443.228ZM458.876 380.217C459.548 380.217 460.14 380.105 460.652 379.881C461.164 379.657 461.564 379.337 461.852 378.921C462.172 378.505 462.332 378.009 462.332 377.433C462.332 376.569 462.012 375.881 461.372 375.369C460.764 374.857 459.932 374.601 458.876 374.601H454.22V380.217H458.876ZM458.828 367.833C459.436 367.833 459.964 367.721 460.412 367.497C460.892 367.273 461.26 366.969 461.516 366.585C461.772 366.201 461.9 365.753 461.9 365.241C461.9 364.473 461.612 363.849 461.036 363.369C460.492 362.889 459.756 362.649 458.828 362.649H454.22V367.833H458.828ZM477.353 388.233V354.633H488.969V388.233H477.353ZM494.275 388.233V354.633H505.891V379.353H521.731V388.233H494.275ZM524.978 388.233V354.633H536.594V388.233H524.978ZM549.82 388.233V362.745H539.836V354.633H571.42V362.745H561.388V388.233H549.82ZM583.548 388.233V375.849L571.548 354.633H583.836L585.66 358.185C586.652 360.137 587.276 362.105 587.532 364.089C587.82 366.041 587.932 368.313 587.868 370.905H591.42C591.356 368.441 591.5 366.121 591.852 363.945C592.204 361.737 592.86 359.641 593.82 357.657L595.404 354.633H607.164L595.116 375.849V388.233H583.548Z" fill="white" />
-                  <path d="M323.373 337L334.013 309H342.173L352.693 337H345.773L339.853 319.56C339.56 318.68 339.36 317.72 339.253 316.68C339.173 315.613 339.133 314.2 339.133 312.44H336.893C336.893 314.28 336.84 315.707 336.733 316.72C336.653 317.733 336.453 318.693 336.133 319.6L330.093 337H323.373ZM329.773 330.52L331.533 325.4H344.493L346.253 330.52H329.773ZM356.141 337V309H368.821C370.821 309 372.581 309.373 374.101 310.12C375.648 310.84 376.848 311.893 377.701 313.28C378.555 314.667 378.981 316.267 378.981 318.08C378.981 319.947 378.555 321.573 377.701 322.96C376.848 324.32 375.648 325.373 374.101 326.12C372.581 326.867 370.835 327.24 368.861 327.24H362.421V337H356.141ZM367.701 322.12C368.635 322.12 369.461 321.96 370.181 321.64C370.901 321.32 371.461 320.867 371.861 320.28C372.288 319.667 372.501 318.933 372.501 318.08C372.501 317.307 372.288 316.627 371.861 316.04C371.461 315.427 370.901 314.96 370.181 314.64C369.461 314.32 368.635 314.16 367.701 314.16H362.421V322.12H367.701ZM383.173 337V309H395.853C397.853 309 399.613 309.373 401.133 310.12C402.679 310.84 403.879 311.893 404.733 313.28C405.586 314.667 406.013 316.267 406.013 318.08C406.013 319.947 405.586 321.573 404.733 322.96C403.879 324.32 402.679 325.373 401.133 326.12C399.613 326.867 397.866 327.24 395.893 327.24H389.453V337H383.173ZM394.733 322.12C395.666 322.12 396.493 321.96 397.213 321.64C397.933 321.32 398.493 320.867 398.893 320.28C399.319 319.667 399.533 318.933 399.533 318.08C399.533 317.307 399.319 316.627 398.893 316.04C398.493 315.427 397.933 314.96 397.213 314.64C396.493 314.32 395.666 314.16 394.733 314.16H389.453V322.12H394.733ZM409.444 324.88V319.6H423.844V324.88H409.444ZM429.657 337V309H442.337C444.337 309 446.097 309.373 447.617 310.12C449.164 310.84 450.364 311.893 451.217 313.28C452.07 314.667 452.497 316.267 452.497 318.08C452.497 319.947 452.07 321.573 451.217 322.96C450.364 324.32 449.164 325.373 447.617 326.12C446.097 326.867 444.35 327.24 442.377 327.24H435.937V337H429.657ZM441.217 322.12C442.15 322.12 442.977 321.96 443.697 321.64C444.417 321.32 444.977 320.867 445.377 320.28C445.804 319.667 446.017 318.933 446.017 318.08C446.017 317.307 445.804 316.627 445.377 316.04C444.977 315.427 444.417 314.96 443.697 314.64C442.977 314.32 442.15 314.16 441.217 314.16H435.937V322.12H441.217ZM456.688 337V309H470.648C472.648 309 474.368 309.307 475.808 309.92C477.248 310.507 478.342 311.36 479.088 312.48C479.862 313.6 480.248 314.893 480.248 316.36C480.248 318.2 479.795 319.587 478.888 320.52C478.008 321.453 476.955 322.067 475.728 322.36C474.528 322.653 473.128 322.84 471.528 322.92V325C473.208 325.027 474.582 325.187 475.648 325.48C476.715 325.773 477.582 326.307 478.248 327.08C478.915 327.827 479.368 328.92 479.608 330.36L480.768 337H474.088L473.208 330.76C473.048 329.693 472.768 328.853 472.368 328.24C471.995 327.6 471.448 327.133 470.728 326.84C470.008 326.547 469.075 326.4 467.928 326.4H462.968V337H456.688ZM469.448 321.4C470.755 321.4 471.795 321.093 472.568 320.48C473.368 319.867 473.768 318.96 473.768 317.76C473.768 316.587 473.368 315.693 472.568 315.08C471.795 314.467 470.755 314.16 469.448 314.16H462.968V321.4H469.448ZM498.916 337.64C495.796 337.64 493.103 337.027 490.836 335.8C488.57 334.547 486.836 332.813 485.636 330.6C484.463 328.387 483.876 325.88 483.876 323.08C483.876 320.28 484.476 317.773 485.676 315.56C486.903 313.32 488.636 311.56 490.876 310.28C493.143 309 495.823 308.36 498.916 308.36C501.983 308.36 504.636 309 506.876 310.28C509.143 311.56 510.876 313.32 512.076 315.56C513.303 317.773 513.916 320.28 513.916 323.08C513.916 325.88 513.316 328.387 512.116 330.6C510.943 332.813 509.223 334.547 506.956 335.8C504.69 337.027 502.01 337.64 498.916 337.64ZM498.956 332.48C500.663 332.48 502.156 332.08 503.436 331.28C504.716 330.453 505.69 329.333 506.356 327.92C507.023 326.48 507.356 324.853 507.356 323.04C507.356 321.2 507.023 319.56 506.356 318.12C505.69 316.68 504.716 315.56 503.436 314.76C502.156 313.933 500.663 313.52 498.956 313.52C497.196 313.52 495.676 313.933 494.396 314.76C493.116 315.56 492.143 316.68 491.476 318.12C490.81 319.56 490.476 321.2 490.476 323.04C490.476 324.853 490.81 326.48 491.476 327.92C492.143 329.333 493.116 330.453 494.396 331.28C495.676 332.08 497.196 332.48 498.956 332.48ZM524.974 337L515.294 309H522.294L526.894 323.92C527.4 325.6 527.76 327.107 527.974 328.44C528.187 329.747 528.294 331.133 528.294 332.6H530.534C530.534 331.16 530.64 329.787 530.854 328.48C531.067 327.147 531.44 325.627 531.974 323.92L536.614 309H543.494L533.854 337H524.974ZM547.235 337V309H568.475V314.16H553.515V320.08H564.875V325.12H553.515V331.84H568.475V337H547.235Z" fill="white" />
-                </g>
-                <defs>
-                  <filter id="filter0_d_421_1186" x="-27.4332" y="-29.7442" width="949.233" height="949.233" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
-                    <feFlood floodOpacity="0" result="BackgroundImageFix" />
-                    <feColorMatrix in="SourceAlpha" type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha" />
-                    <feOffset dy="77.4884" />
-                    <feGaussianBlur stdDeviation="58.1163" />
-                    <feComposite in2="hardAlpha" operator="out" />
-                    <feColorMatrix type="matrix" values="0 0 0 0 0.258824 0 0 0 0 0.521569 0 0 0 0 0.956863 0 0 0 0.2 0" />
-                    <feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow_421_1186" />
-                    <feBlend mode="normal" in="SourceGraphic" in2="effect1_dropShadow_421_1186" result="shape" />
-                  </filter>
-                  <linearGradient id="paint0_linear_421_1186" x1="447.183" y1="9.00021" x2="447.183" y2="725.768" gradientUnits="userSpaceOnUse">
-                    <stop stopColor="#4285F4" />
-                    <stop offset="1" stopColor="#033D9C" />
-                  </linearGradient>
-                  <linearGradient id="paint1_linear_421_1186" x1="-34.2943" y1="-46.9972" x2="510.187" y2="924.853" gradientUnits="userSpaceOnUse">
-                    <stop stopColor="white" />
-                    <stop offset="1" stopColor="white" stopOpacity="0" />
-                  </linearGradient>
-                  <linearGradient id="paint2_linear_421_1186" x1="447.183" y1="125.233" x2="447.183" y2="628.907" gradientUnits="userSpaceOnUse">
-                    <stop stopColor="white" />
-                    <stop offset="0.0001" stopColor="white" stopOpacity="0.81" />
-                    <stop offset="1" stopColor="white" stopOpacity="0" />
-                  </linearGradient>
-                </defs>
-              </svg>
-            </div>
+          <SecurityCertification></SecurityCertification>
+          <ReliabilityCertification></ReliabilityCertification>
           </div>
         </div>
         <Separator />
