+++ app/publish/page.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 
 import { redirect } from 'next/navigation'
-import PublishForm from "@/components/publish-form"
+import PublishForm from "@/components/Publish/Form"
 import { createClient } from "@/utils/supabase/server"
 import { CheckoutForm } from "@/components/checkout"
