+++ components/Certifications/ReliabilityCertification.tsx
@@ -1,0 +1,42 @@
+export default function ReliabilityCertification() {
+    return (
+        <div className="flex w-36 h-36 sm:w-64 sm:h-64">
+            <svg viewBox="0 0 826 746" fill="none" xmlns="http://www.w3.org/2000/svg" className=" overflow-visible">
+                <g filter="url(#filter0_d_421_1186)">
+                    <path d="M397.894 60.3472C424.783 32.3351 469.583 32.3351 496.472 60.3472L501.772 65.8681C518.514 83.3094 543.228 90.604 566.758 85.0497L574.074 83.3227C611.708 74.439 649.208 98.5309 656.773 136.453L658.535 145.283C663.223 168.786 679.883 188.113 702.437 196.216L710.673 199.176C746.788 212.151 765.137 252.317 751.3 288.11L747.7 297.422C739.121 319.614 742.705 344.67 757.16 363.568L763.182 371.44C786.446 401.853 780.177 445.446 749.284 468.071L741.659 473.656C722.412 487.752 711.911 510.867 713.957 534.635L714.784 544.246C718.087 582.62 689.035 616.138 650.581 618.318L642.563 618.773C618.516 620.136 596.964 634.06 585.837 655.42L582.02 662.746C564.101 697.145 521.174 709.744 487.506 690.488L481.103 686.827C460.087 674.807 434.279 674.807 413.263 686.827L406.86 690.488C373.192 709.744 330.265 697.145 312.346 662.746L308.529 655.42C297.402 634.06 275.85 620.136 251.803 618.773L243.785 618.318C205.331 616.138 176.279 582.62 179.582 544.246L180.409 534.635C182.455 510.867 171.954 487.752 152.707 473.656L145.082 468.071C114.189 445.446 107.92 401.853 131.185 371.44L137.206 363.568C151.661 344.67 155.245 319.614 146.666 297.422L143.066 288.11C129.229 252.317 147.578 212.151 183.693 199.176L191.929 196.216C214.484 188.113 231.143 168.785 235.832 145.283L237.593 136.453C245.158 98.5309 282.658 74.439 320.292 83.3227L327.608 85.0497C351.138 90.604 375.852 83.3094 392.594 65.868L397.894 60.3472Z" fill="url(#paint0_linear_421_1186)" />
+                    <path d="M397.894 60.3472C424.783 32.3351 469.583 32.3351 496.472 60.3472L501.772 65.8681C518.514 83.3094 543.228 90.604 566.758 85.0497L574.074 83.3227C611.708 74.439 649.208 98.5309 656.773 136.453L658.535 145.283C663.223 168.786 679.883 188.113 702.437 196.216L710.673 199.176C746.788 212.151 765.137 252.317 751.3 288.11L747.7 297.422C739.121 319.614 742.705 344.67 757.16 363.568L763.182 371.44C786.446 401.853 780.177 445.446 749.284 468.071L741.659 473.656C722.412 487.752 711.911 510.867 713.957 534.635L714.784 544.246C718.087 582.62 689.035 616.138 650.581 618.318L642.563 618.773C618.516 620.136 596.964 634.06 585.837 655.42L582.02 662.746C564.101 697.145 521.174 709.744 487.506 690.488L481.103 686.827C460.087 674.807 434.279 674.807 413.263 686.827L406.86 690.488C373.192 709.744 330.265 697.145 312.346 662.746L308.529 655.42C297.402 634.06 275.85 620.136 251.803 618.773L243.785 618.318C205.331 616.138 176.279 582.62 179.582 544.246L180.409 534.635C182.455 510.867 171.954 487.752 152.707 473.656L145.082 468.071C114.189 445.446 107.92 401.853 131.185 371.44L137.206 363.568C151.661 344.67 155.245 319.614 146.666 297.422L143.066 288.11C129.229 252.317 147.578 212.151 183.693 199.176L191.929 196.216C214.484 188.113 231.143 168.785 235.832 145.283L237.593 136.453C245.158 98.5309 282.658 74.439 320.292 83.3227L327.608 85.0497C351.138 90.604 375.852 83.3094 392.594 65.868L397.894 60.3472Z" fill="url(#paint1_linear_421_1186)" fillOpacity="0.1" />
+                    <path d="M198.702 372.236C198.702 235.005 309.95 123.757 447.181 123.757C584.412 123.757 695.66 235.005 695.66 372.236V377.346C695.66 514.577 584.412 625.825 447.181 625.825C309.95 625.825 198.702 514.577 198.702 377.346V372.236Z" fill="black" fillOpacity="0.04" />
+                    <path d="M224.212 372.236C224.212 249.094 324.039 149.267 447.181 149.267C570.323 149.267 670.149 249.094 670.149 372.236V377.346C670.149 500.489 570.323 600.315 447.181 600.315C324.039 600.315 224.212 500.488 224.212 377.346V372.236Z" stroke="white" strokeOpacity="0.3" strokeWidth="51.0205" />
+                    <path d="M699.02 377.07C699.02 516.156 586.269 628.907 447.183 628.907C308.097 628.907 195.346 516.156 195.346 377.07C195.346 237.984 308.097 125.233 447.183 125.233C586.269 125.233 699.02 237.984 699.02 377.07ZM247.193 377.07C247.193 487.521 336.731 577.06 447.183 577.06C557.634 577.06 647.173 487.521 647.173 377.07C647.173 266.618 557.634 177.08 447.183 177.08C336.731 177.08 247.193 266.618 247.193 377.07Z" fill="url(#paint2_linear_421_1186)" fillOpacity="0.5" />
+                    <path d="M290.415 388.233V354.633H309.663C312.255 354.633 314.463 354.985 316.287 355.689C318.143 356.393 319.535 357.401 320.463 358.713C321.391 359.993 321.855 361.481 321.855 363.177C321.855 365.289 321.247 366.921 320.031 368.073C318.847 369.193 317.375 369.945 315.615 370.329C313.887 370.713 311.951 370.905 309.807 370.905V374.265C312.335 374.265 314.351 374.489 315.855 374.937C317.359 375.353 318.559 376.105 319.455 377.193C320.383 378.281 321.103 379.849 321.615 381.897L323.151 388.233H311.103L309.423 380.073C309.231 379.145 308.991 378.425 308.703 377.913C308.415 377.369 308.015 376.985 307.503 376.761C307.023 376.505 306.367 376.377 305.535 376.377H301.647V388.233H290.415ZM306.639 369.081C307.727 369.081 308.591 368.809 309.231 368.265C309.903 367.689 310.239 366.953 310.239 366.057C310.239 365.129 309.919 364.393 309.279 363.849C308.639 363.305 307.759 363.033 306.639 363.033H301.647V369.081H306.639ZM326.697 388.233V354.633H353.625V363.225H337.737V367.401H350.457V375.129H337.737V379.593H353.625V388.233H326.697ZM358.103 388.233V354.633H369.719V379.353H385.559V388.233H358.103ZM388.806 388.233V354.633H400.422V388.233H388.806ZM402.848 388.233L414.08 354.633H429.584L440.912 388.233H428.672L424.208 372.441C423.792 370.841 423.536 369.017 423.44 366.969C423.376 364.921 423.392 362.569 423.488 359.913H419.84C419.936 362.313 419.904 364.585 419.744 366.729C419.584 368.873 419.264 370.777 418.784 372.441L414.128 388.233H402.848ZM410.912 382.473L413.84 374.361H429.392L432.368 382.473H410.912ZM443.228 388.233V354.633H460.94C463.34 354.633 465.388 354.937 467.084 355.545C468.812 356.153 470.108 357.017 470.972 358.137C471.836 359.257 472.268 360.537 472.268 361.977C472.268 364.249 471.452 366.009 469.82 367.257C468.188 368.473 465.74 369.113 462.476 369.177V372.489C464.908 372.489 466.94 372.761 468.572 373.305C470.236 373.849 471.484 374.697 472.316 375.849C473.148 376.969 473.564 378.393 473.564 380.121C473.564 381.817 473.068 383.273 472.076 384.489C471.116 385.705 469.692 386.633 467.804 387.273C465.948 387.913 463.724 388.233 461.132 388.233H443.228ZM458.876 380.217C459.548 380.217 460.14 380.105 460.652 379.881C461.164 379.657 461.564 379.337 461.852 378.921C462.172 378.505 462.332 378.009 462.332 377.433C462.332 376.569 462.012 375.881 461.372 375.369C460.764 374.857 459.932 374.601 458.876 374.601H454.22V380.217H458.876ZM458.828 367.833C459.436 367.833 459.964 367.721 460.412 367.497C460.892 367.273 461.26 366.969 461.516 366.585C461.772 366.201 461.9 365.753 461.9 365.241C461.9 364.473 461.612 363.849 461.036 363.369C460.492 362.889 459.756 362.649 458.828 362.649H454.22V367.833H458.828ZM477.353 388.233V354.633H488.969V388.233H477.353ZM494.275 388.233V354.633H505.891V379.353H521.731V388.233H494.275ZM524.978 388.233V354.633H536.594V388.233H524.978ZM549.82 388.233V362.745H539.836V354.633H571.42V362.745H561.388V388.233H549.82ZM583.548 388.233V375.849L571.548 354.633H583.836L585.66 358.185C586.652 360.137 587.276 362.105 587.532 364.089C587.82 366.041 587.932 368.313 587.868 370.905H591.42C591.356 368.441 591.5 366.121 591.852 363.945C592.204 361.737 592.86 359.641 593.82 357.657L595.404 354.633H607.164L595.116 375.849V388.233H583.548Z" fill="white" />
+                    <path d="M323.373 337L334.013 309H342.173L352.693 337H345.773L339.853 319.56C339.56 318.68 339.36 317.72 339.253 316.68C339.173 315.613 339.133 314.2 339.133 312.44H336.893C336.893 314.28 336.84 315.707 336.733 316.72C336.653 317.733 336.453 318.693 336.133 319.6L330.093 337H323.373ZM329.773 330.52L331.533 325.4H344.493L346.253 330.52H329.773ZM356.141 337V309H368.821C370.821 309 372.581 309.373 374.101 310.12C375.648 310.84 376.848 311.893 377.701 313.28C378.555 314.667 378.981 316.267 378.981 318.08C378.981 319.947 378.555 321.573 377.701 322.96C376.848 324.32 375.648 325.373 374.101 326.12C372.581 326.867 370.835 327.24 368.861 327.24H362.421V337H356.141ZM367.701 322.12C368.635 322.12 369.461 321.96 370.181 321.64C370.901 321.32 371.461 320.867 371.861 320.28C372.288 319.667 372.501 318.933 372.501 318.08C372.501 317.307 372.288 316.627 371.861 316.04C371.461 315.427 370.901 314.96 370.181 314.64C369.461 314.32 368.635 314.16 367.701 314.16H362.421V322.12H367.701ZM383.173 337V309H395.853C397.853 309 399.613 309.373 401.133 310.12C402.679 310.84 403.879 311.893 404.733 313.28C405.586 314.667 406.013 316.267 406.013 318.08C406.013 319.947 405.586 321.573 404.733 322.96C403.879 324.32 402.679 325.373 401.133 326.12C399.613 326.867 397.866 327.24 395.893 327.24H389.453V337H383.173ZM394.733 322.12C395.666 322.12 396.493 321.96 397.213 321.64C397.933 321.32 398.493 320.867 398.893 320.28C399.319 319.667 399.533 318.933 399.533 318.08C399.533 317.307 399.319 316.627 398.893 316.04C398.493 315.427 397.933 314.96 397.213 314.64C396.493 314.32 395.666 314.16 394.733 314.16H389.453V322.12H394.733ZM409.444 324.88V319.6H423.844V324.88H409.444ZM429.657 337V309H442.337C444.337 309 446.097 309.373 447.617 310.12C449.164 310.84 450.364 311.893 451.217 313.28C452.07 314.667 452.497 316.267 452.497 318.08C452.497 319.947 452.07 321.573 451.217 322.96C450.364 324.32 449.164 325.373 447.617 326.12C446.097 326.867 444.35 327.24 442.377 327.24H435.937V337H429.657ZM441.217 322.12C442.15 322.12 442.977 321.96 443.697 321.64C444.417 321.32 444.977 320.867 445.377 320.28C445.804 319.667 446.017 318.933 446.017 318.08C446.017 317.307 445.804 316.627 445.377 316.04C444.977 315.427 444.417 314.96 443.697 314.64C442.977 314.32 442.15 314.16 441.217 314.16H435.937V322.12H441.217ZM456.688 337V309H470.648C472.648 309 474.368 309.307 475.808 309.92C477.248 310.507 478.342 311.36 479.088 312.48C479.862 313.6 480.248 314.893 480.248 316.36C480.248 318.2 479.795 319.587 478.888 320.52C478.008 321.453 476.955 322.067 475.728 322.36C474.528 322.653 473.128 322.84 471.528 322.92V325C473.208 325.027 474.582 325.187 475.648 325.48C476.715 325.773 477.582 326.307 478.248 327.08C478.915 327.827 479.368 328.92 479.608 330.36L480.768 337H474.088L473.208 330.76C473.048 329.693 472.768 328.853 472.368 328.24C471.995 327.6 471.448 327.133 470.728 326.84C470.008 326.547 469.075 326.4 467.928 326.4H462.968V337H456.688ZM469.448 321.4C470.755 321.4 471.795 321.093 472.568 320.48C473.368 319.867 473.768 318.96 473.768 317.76C473.768 316.587 473.368 315.693 472.568 315.08C471.795 314.467 470.755 314.16 469.448 314.16H462.968V321.4H469.448ZM498.916 337.64C495.796 337.64 493.103 337.027 490.836 335.8C488.57 334.547 486.836 332.813 485.636 330.6C484.463 328.387 483.876 325.88 483.876 323.08C483.876 320.28 484.476 317.773 485.676 315.56C486.903 313.32 488.636 311.56 490.876 310.28C493.143 309 495.823 308.36 498.916 308.36C501.983 308.36 504.636 309 506.876 310.28C509.143 311.56 510.876 313.32 512.076 315.56C513.303 317.773 513.916 320.28 513.916 323.08C513.916 325.88 513.316 328.387 512.116 330.6C510.943 332.813 509.223 334.547 506.956 335.8C504.69 337.027 502.01 337.64 498.916 337.64ZM498.956 332.48C500.663 332.48 502.156 332.08 503.436 331.28C504.716 330.453 505.69 329.333 506.356 327.92C507.023 326.48 507.356 324.853 507.356 323.04C507.356 321.2 507.023 319.56 506.356 318.12C505.69 316.68 504.716 315.56 503.436 314.76C502.156 313.933 500.663 313.52 498.956 313.52C497.196 313.52 495.676 313.933 494.396 314.76C493.116 315.56 492.143 316.68 491.476 318.12C490.81 319.56 490.476 321.2 490.476 323.04C490.476 324.853 490.81 326.48 491.476 327.92C492.143 329.333 493.116 330.453 494.396 331.28C495.676 332.08 497.196 332.48 498.956 332.48ZM524.974 337L515.294 309H522.294L526.894 323.92C527.4 325.6 527.76 327.107 527.974 328.44C528.187 329.747 528.294 331.133 528.294 332.6H530.534C530.534 331.16 530.64 329.787 530.854 328.48C531.067 327.147 531.44 325.627 531.974 323.92L536.614 309H543.494L533.854 337H524.974ZM547.235 337V309H568.475V314.16H553.515V320.08H564.875V325.12H553.515V331.84H568.475V337H547.235Z" fill="white" />
+                </g>
+                <defs>
+                    <filter id="filter0_d_421_1186" x="-27.4332" y="-29.7442" width="949.233" height="949.233" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
+                        <feFlood floodOpacity="0" result="BackgroundImageFix" />
+                        <feColorMatrix in="SourceAlpha" type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha" />
+                        <feOffset dy="77.4884" />
+                        <feGaussianBlur stdDeviation="58.1163" />
+                        <feComposite in2="hardAlpha" operator="out" />
+                        <feColorMatrix type="matrix" values="0 0 0 0 0.258824 0 0 0 0 0.521569 0 0 0 0 0.956863 0 0 0 0.2 0" />
+                        <feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow_421_1186" />
+                        <feBlend mode="normal" in="SourceGraphic" in2="effect1_dropShadow_421_1186" result="shape" />
+                    </filter>
+                    <linearGradient id="paint0_linear_421_1186" x1="447.183" y1="9.00021" x2="447.183" y2="725.768" gradientUnits="userSpaceOnUse">
+                        <stop stopColor="#4285F4" />
+                        <stop offset="1" stopColor="#033D9C" />
+                    </linearGradient>
+                    <linearGradient id="paint1_linear_421_1186" x1="-34.2943" y1="-46.9972" x2="510.187" y2="924.853" gradientUnits="userSpaceOnUse">
+                        <stop stopColor="white" />
+                        <stop offset="1" stopColor="white" stopOpacity="0" />
+                    </linearGradient>
+                    <linearGradient id="paint2_linear_421_1186" x1="447.183" y1="125.233" x2="447.183" y2="628.907" gradientUnits="userSpaceOnUse">
+                        <stop stopColor="white" />
+                        <stop offset="0.0001" stopColor="white" stopOpacity="0.81" />
+                        <stop offset="1" stopColor="white" stopOpacity="0" />
+                    </linearGradient>
+                </defs>
+            </svg>
+        </div>
+    )
+}
