+++ components/Offers/OffersPagination.tsx
