+++ components/Offers/OffersSearch.tsx
