+++ components/Offers/OffersSkeleton.tsx
@@ -1,4 +1,4 @@
-import { Skeleton } from "./ui/skeleton"
+import { Skeleton } from "../ui/skeleton"
 
 export default function OffersSkeleton() {
     return (
